test(#1646): positionTrainPriority test cases backendSsotCascade로 통합 (SonarCloud dup)
- 별도 파일(useFusedNearestStation.positionTrainPriority.test.ts) 삭제
- backendSsotCascade.test.ts에 #1646 describe block 추가 (8 케이스)
- jest.mock 셋업 중복 제거로 SonarCloud new_duplicated_lines_density 8.2% → 미만 목표
- 주석에서 '3-of-3 합의'와 prerequisite(positionTrainResult) 구분 명확화
